+++ CHANGES.rst
@@ -4,6 +4,8 @@ Changelog (niondata)
 UNRELEASED
 ----------
 
+- Allow persistent items to be hidden (like properties).
+
 - Allow FilteredListModel to have separate master/item property names.
 
 0.3.12 (2018-07-23)